+++ test/eclipse/src/lombok/eclipse/EclipseTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022-2024 The Project Lombok Authors.
+ * Copyright (C) 2022-2025 The Project Lombok Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -27,6 +27,7 @@ import org.junit.runners.Suite.SuiteClasses;
 
 import lombok.eclipse.cleanup.CleanupTest;
 import lombok.eclipse.compile.NoErrorsTest;
+import lombok.eclipse.edit.FoldingTest;
 import lombok.eclipse.edit.SelectTest;
 import lombok.eclipse.misc.DelegateTest;
 import lombok.eclipse.misc.JavadocTest;
@@ -37,7 +38,7 @@ import lombok.eclipse.refactoring.RenameTest;
 import lombok.eclipse.references.FindReferencesTest;
 
 @RunWith(Suite.class)
-@SuiteClasses({ExtractInterfaceTest.class, RenameTest.class, SelectTest.class, CleanupTest.class, FindReferencesTest.class, InlineTest.class, NoErrorsTest.class, JavadocTest.class, DelegateTest.class, ExtractVariableTest.class})
+@SuiteClasses({ExtractInterfaceTest.class, RenameTest.class, SelectTest.class, CleanupTest.class, FindReferencesTest.class, InlineTest.class, NoErrorsTest.class, JavadocTest.class, DelegateTest.class, ExtractVariableTest.class, FoldingTest.class})
 public class EclipseTests {
 	
 }